% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abmAME-package.R
\docType{package}
\name{abmAME-package}
\alias{abmAME}
\alias{abmAME-package}
\title{abmAME: An R Package for Simulating Animal Movement Using an Agent-Based Model EXTENDED}
\description{
The *abmAME* simulates animal movement use a discrete time agent-based model, programmed in C++ via the Rcpp package. The simulations include a number of key internal and external movement influences, as well as parameters for navigation and mobility capacity of the animal.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/margaret-swift/abmAME}
  \item Report bugs at \url{https://github.com/margaret-swift/abmAME}
}

}
\author{
\strong{Maintainer}: Margaret Elizabeth Swift \email{margaret.swift@cornell.edu} (\href{https://orcid.org/0000-0001-7913-1879}{ORCID})

}
\keyword{internal}
