% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_intersection}
\alias{check_intersection}
\title{Check if a line segment (step option) crosses a barrier}
\arguments{
\item{origin}{start location}

\item{target}{target location}

\item{barrier_x1}{barrier segment x start values}

\item{barrier_x2}{barrier segment x end values}

\item{barrier_y1}{barrier segment y start values}

\item{barrier_y2}{barrier segment y end values}

\item{p_cross}{barrier segment permeability values}
}
\value{
A Boolean of whether any barrier is crossed
Useful resource: https://www.geeksforgeeks.org/check-if-two-given-line-segments-intersect/
}
\description{
Check if a line segment (step option) crosses a barrier
}
