% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_get_values_rast}
\alias{cpp_get_values_rast}
\title{cpp_get_values_rast}
\arguments{
\item{RASTER}{A RASTER that the value will be extracted from.}

\item{XLOCS}{A vector of x locations.}

\item{YLOCS}{A vector of y locations.}
}
\value{
A vector of values equal to the length of XLOCS extracted from
MATRIX.
}
\description{
cpp_get_values_rast
}
\details{
A simple C++ function to extract a value from a RASTER
using two vectors describing location. Note that the counting for x and y
begins at zero.
}
