% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBarriers.R
\name{generateBarriers}
\alias{generateBarriers}
\title{Generate barriers from sf file and permeability given by user}
\usage{
generateBarriers(barriers, rast, p_list, ...)
}
\arguments{
\item{barriers}{Barriers provided by user in sf format.}

\item{p_list}{Either a list of permeability values (that should correspond
to the length of barriers) or a single value (to be applied to all barriers)}
}
\value{
a matrix of six columns:
}
\description{
This R function generates a barrier dataset from any input
barrier shapefile and user-defined list of corresponding permeability values
(or one value to be applied to the entire list)
}
