% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLandscapeData.R
\name{generateLandscapeData}
\alias{generateLandscapeData}
\title{Generate barriers and landscape data matrices from sf file and permeability given by user}
\usage{
generateLandscapeData(rasters, barriers, perms = 1, crs = 32735)
}
\arguments{
\item{rasters}{Landscape raster data}

\item{barriers}{Barriers provided by user in sf format.}

\item{crs}{a CRS for reprojection}

\item{p_list}{Either a list of permeability values (that should correspond
to the length of barriers) or a single value (to be applied to all barriers)}
}
\value{
a list with: landscape data matrices, landscape data extent and resolution, and barrier data frame
}
\description{
This R function generates a barrier dataset from any input
barrier shapefile and user-defined list of corresponding permeability values
(or one value to be applied to the entire list)
}
