% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_intersection}
\alias{check_intersection}
\title{Check if a line segment (step option) crosses a fence}
\arguments{
\item{origin}{start location}

\item{target}{target location}

\item{fence_x1}{fence segment x start values}

\item{fence_x2}{fence segment x end values}

\item{fence_y1}{fence segment y start values}

\item{fence_y2}{fence segment y end values}

\item{p_cross}{fence segment permeability values}
}
\value{
A Boolean of whether any fence is crossed
Useful resource: https://www.geeksforgeeks.org/check-if-two-given-line-segments-intersect/
}
\description{
Check if a line segment (step option) crosses a fence
}
